refactor(webhook): guard OVMS auto-versioning behind distro build tag
Moves OVMS auto-versioning init container injection into build-tag-guarded
companion files so the OCP-specific code only compiles into the distro
build. The upstream (non-distro) build gets a no-op stub, keeping the
shared codebase clean and the feature invisible to upstream consumers.

Introduces an injectPlatformExtensions dispatch point in the storage
initializer injector, providing a clean hook for future platform-specific
pod mutations without touching the main reconciliation path.
