+++ pkg/webhook/admission/pod/storage_initializer_injector_default.go
@@ -1,0 +1,27 @@
+//go:build !distro
+
+/*
+Copyright 2021 The KServe Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package pod
+
+import corev1 "k8s.io/api/core/v1"
+
+// injectPlatformExtensions is a hook for distribution-specific pod mutations
+// after storage initialization (e.g. OVMS auto-versioning on OCP).
+func (mi *StorageInitializerInjector) injectPlatformExtensions(_ *corev1.Pod) error {
+	return nil
+}
